being explicit about CK_ULONG and CK_LONG types

+++ src/types.rs
@@ -65,14 +65,14 @@ pub type CK_BBOOL = CK_BYTE;
 #[cfg(windows)]
 pub type CK_ULONG = u32;
 #[cfg(not(windows))]
-pub type CK_ULONG = usize;
+pub type CK_ULONG = u64;
 pub type CK_ULONG_PTR = *mut CK_ULONG;
 
 /// a signed value, the same size as a CK_ULONG
 #[cfg(windows)]
 pub type CK_LONG = i32;
 #[cfg(not(windows))]
-pub type CK_LONG = isize;
+pub type CK_LONG = i64;
 
 
 /// at least 32 bits; each bit is a Boolean flag
